move some logic out of fayserver.js
